+++ controllers/contacts.js
@@ -1,14 +1,8 @@
-const Joi = require("joi");
-
 const contacts = require("../models/contacts");
 
-const { HttpError, ctrlWrapper } = require("../helpers");
+const HttpError = require("../helpers");
 
-const addSchema = Joi.object({
-    name: Joi.string().required(),
-    email: Joi.string().required(),
-    phone: Joi.string().required(),
-});
+const { ctrlWrapper } = require("../decorators");
 
 const getAll = async (req, res, next) => {
             const result = await contacts.listContacts();
@@ -25,11 +19,7 @@ const getById = async (req, res, next) => {
     };
 
 const createContact = async (req, res, next) => {
-        const { error } = addSchema.validate(req.body)
-        if (error) {
-            throw HttpError(400, error.message)
-        }
-    const result = await contacts.addContact(req.body);
+        const result = await contacts.addContact(req.body);
         res.status(201).json(result)
     };
 
@@ -45,20 +35,12 @@ const deleteById = async (req, res, next) => {
     };
 
 const updateContactsById = async (req, res, next) => {
-    const { error } = addSchema.validate(req.body)
-    if (error) {
-      throw HttpError(400, error.message)
-    };
     const { id } = req.params;
     const result = await contacts.updateContact(id, req.body);
     if (!result) {
      throw HttpError(404, "Not found")
     }
-    const response = {
-      message: "Your contact was updated",
-      result: result
-    };
-    res.json(response)
+    res.json(result)
   }
