+++ decorators/index.js
@@ -1,0 +1,10 @@
+const ctrlWrapper = require("./ctrlWrapper");
+const validateBody = require("./validateBody")
+
+
+
+
+module.exports = {
+    ctrlWrapper,
+    validateBody
+};
