+++ helpers/index.js
@@ -1,7 +1,6 @@
 const HttpError = require("./HttpError")
-const ctrlWrapper = require("./ctrlWrapper")
+
 
 module.exports = {
     HttpError,
-    ctrlWrapper,
-};
+ };
