+++ schemas/contacts.js
@@ -1,0 +1,9 @@
+const Joi = require("joi");
+
+const addSchema = Joi.object({
+    name: Joi.string().required(),
+    email: Joi.string().required(),
+    phone: Joi.string().required(),
+});
+
+module.exports = addSchema;
